rename to trajectory_follower_base, trajectory_follower_node
