+++ control/mpc_lateral_controller/README.md
@@ -1,7 +1,7 @@
 # MPC Lateral Controller
 
 This is the design document for the lateral controller node
-in the `mpc_lateral_controller_nodes` package.
+in the `trajectory_follower_node` package.
 
 ## Purpose / Use cases
 
@@ -58,7 +58,7 @@ The tracking is not accurate if the first point of the reference trajectory is a
 
 ### Inputs
 
-Set the following from the [controller_node](../../mpc_lateral_controller_nodes/design/mpc_lateral_controller-design.md)
+Set the following from the [controller_node](../../trajectory_follower_node/design/mpc_lateral_controller-design.md)
 
 - `autoware_auto_planning_msgs/Trajectory` : reference trajectory to follow.
 - `nav_msgs/Odometry`: current odometry
@@ -66,7 +66,7 @@ Set the following from the [controller_node](../../mpc_lateral_controller_nodes/
 
 ### Outputs
 
-Return LateralOutput which contains the following to [controller_node](../../mpc_lateral_controller_nodes/design/mpc_lateral_controller-design.md)
+Return LateralOutput which contains the following to [controller_node](../../trajectory_follower_node/design/mpc_lateral_controller-design.md)
 
 - `autoware_auto_control_msgs/AckermannLateralCommand`
 - LateralSyncData
