+++ control/pid_longitudinal_controller/package.xml
@@ -35,7 +35,7 @@
   <depend>tf2_ros</depend>
   <depend>tier4_autoware_utils</depend>
   <depend>tier4_debug_msgs</depend>
-  <depend>trajectory_follower</depend>
+  <depend>trajectory_follower_base</depend>
   <depend>vehicle_info_util</depend>
 
   <test_depend>ament_cmake_ros</test_depend>
