+++ control/trajectory_follower_base/package.xml
@@ -1,11 +1,13 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>mpc_lateral_controller</name>
+  <name>trajectory_follower_base</name>
   <version>1.0.0</version>
-  <description>MPC-based lateral controller</description>
+  <description>Library for generating lateral and longitudinal controls following a trajectory</description>
 
+  <!-- mpc longitudinal controller -->
   <maintainer email="takamasa.horibe@tier4.jp">Takamasa Horibe</maintainer>
+  <!-- pid lateral controller -->
   <maintainer email="takayuki.murooka@tier4.jp">Takayuki Murooka</maintainer>
 
   <license>Apache 2.0</license>
@@ -36,7 +38,6 @@
   <depend>tf2_ros</depend>
   <depend>tier4_autoware_utils</depend>
   <depend>tier4_debug_msgs</depend>
-  <depend>trajectory_follower_base</depend>
   <depend>vehicle_info_util</depend>
 
   <test_depend>ament_cmake_ros</test_depend>
