+++ control/trajectory_follower_node/design/simple_trajectory_follower-design.md
