+++ launch/tier4_control_launch/package.xml
@@ -15,8 +15,7 @@
   <exec_depend>external_cmd_selector</exec_depend>
   <exec_depend>lane_departure_checker</exec_depend>
   <exec_depend>shift_decider</exec_depend>
-  <exec_depend>trajectory_follower</exec_depend>
-  <exec_depend>trajectory_follower_nodes</exec_depend>
+  <exec_depend>trajectory_follower_node</exec_depend>
   <exec_depend>vehicle_cmd_gate</exec_depend>
 
   <test_depend>ament_lint_auto</test_depend>
